Changes to routes files are now propagated
Before I used require to load the modules from routes.js and ftlRoutes.js.
Due to the cache of require this resulted in not getting updated versions
when the files changed.
As a solution to this there is now a function which synchronouly reads the
files and creates the modules within them. This way we actually always get
the up-to-date versions of the modules.
Tested this by running puer from the commandline and changing some test
files, works like a charm.