Displaying freemarker based mocked routs works now
Though there are a few problems with static files now. Maybe revert back
to old approach.
Trouble was that 'require' was used to get the js files and require keeps
a cache so after an update to the file a call to require still returned
the old version. this is now solved with 'delete
require.cache[require.resolve(requirePath)];'.